\name{dRUMocc}
\alias{dRUMocc}

\title{dRUMocc - single-season Bayesian occupancy model using the dRUM formulation.}

\description{
The function undertakes MCMC sampling of the parameters of a single-season Bayesian occupancy model that uses the dRUM formulation.
}

\usage{
dRUMocc(formula, design_mats, ndraws, alpha_m, beta_m, sigma_inv_alpha_p, sigma_inv_beta_p)
}
\arguments{
  \item{formula }{ - Double right-hand side formula describing covariates of detection and occupancy in that order.}
  \item{design_mats }{ - A named list generated by the call vb_Designs(W, X, y).}
  \item{ndraws }{ - The number of MCMC samples undertaken. At the moment this is set as the number of burn-in iterations plus the number of post burn-in samples. Half of ndraws are retained. This function is slow relative to PGocc4 and thus SHOULD NOT be used for large simulations! Because of this, the functionality of the function does not allow the user to specify the burn-in proportion.}
  \item{alpha_m }{ - The prior mean alpha vector. At the moment this should be set to a zero column vector.}
  \item{beta_m }{ - The prior mean beta vector. At the moment this should be set to a zero column vector.}
  \item{sigma_inv_alpha_p }{ - The inverse of the covariance matrix of alpha.}
  \item{sigma_inv_beta_p }{ - The inverse of the covariance matrix of beta.}
}
\details{The function undertakes MCMC sampling of the parameters of a single-season Bayesian occupancy model that uses the dRUM formulation. The function does not have much functionality since we advise the use of \code{PGocc} or \code{PGocc2} instead.}
\value{
\item{ }{ - The MCMC samples are returned as a named list with entries 'alpha' and 'beta'. Both are matrices.}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
data("SpatSimData")
#--------------------------------------------------------------------------

Xmat <- SimTable$Xmat[1:805,] #only use surveyed sites (805)
colnames(Xmat) #PC1, PC2

Wmat <- list( W1=SimTable$Wmat )
names(Wmat) #W1

Ysim <- SimTable$Ysim

design_list <- vb_Designs(W=Wmat, X=Xmat, y=Ysim)
names(design_list$Y) # V1
#--------------------------------------------------------------------------

#specification of the prior distributions
beta_m <- matrix(0, nrow=3, ncol=1) #prior mean for beta
sigma_inv_beta <- diag(3)/1000 #prior covariance for beta

alpha_m <- matrix(0, nrow=2, ncol=1) #prior mean for alpha
sigma_inv_alpha <- diag(2)/1000 #prior covariance for alpha

# ~ occupancy covariates ~ detection covariates
#this function can be slow! use PGocc4 instead
occ <- dRUMocc(formula= V1  ~ PC1 + PC2 ~ W1,
            design_mats = design_list,
            ndraws=100,
            alpha_m=alpha_m, beta_m=beta_m,
            sigma_inv_alpha_p=sigma_inv_alpha,
            sigma_inv_beta_p=sigma_inv_beta)

names(occ)
#--------------------------------------------------------------------------
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
