\name{PGocc4}
\alias{PGocc4}

\title{PGocc4 - single-season Bayesian occupancy model using Polya-Gamma random variables.}

\description{
The function undertakes MCMC sampling of the parameters of a single-season Bayesian occupancy model that uses Polya-Gamma random variables.
}

\usage{
PGocc4(formula, design_mats, ndraws, alpha_m, beta_m, sigma_inv_alpha_p, sigma_inv_beta_p)
}
\arguments{
  \item{formula }{ - Double right-hand side formula describing covariates of detection and occupancy in that order.}
  \item{design_mats }{ - A named list generated by the call vb_Designs(W, X, y).}
  \item{ndraws }{ - The number of MCMC samples undertaken. This is set as the number of burn-in iterations plus the number of post burn-in samples. See 'percent_burn_in' below.}
  \item{alpha_m }{ - The prior mean alpha vector. Supply a matrix of the correct dimension.}
  \item{beta_m }{ - The prior mean beta vector. Supply a matrix of the correct dimension.}
  \item{sigma_inv_alpha_p }{ - The alpha precision matrix (i.e. the inverse of the covariance matrix).}
  \item{sigma_inv_beta_p }{ - The beta precision matrix.}
  \item{percent_burn_in }{ - The percentage of 'ndraws' used as burn-in samples. The argument entered should be between 0.0 and 1.0. No check however is done to assess whether the user has entered the correct argument here. }
}
\details{
The function undertakes MCMC sampling of the parameters of a single-season Bayesian occupancy model that uses Polya-Gamma random variables.
}
\value{
\item{ }{  - The MCMC samples are returned as a named list with entries 'alpha' and 'beta'. Both are matrices witht the number of rows equal to the number of regression effects.}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
%% ~~objects to See Also as \code{\link{help}}, ~~~
data("SpatSimData")
#--------------------------------------------------------------------------

Xmat <- SimTable$Xmat[1:805,] #only use surveyed sites (805)
colnames(Xmat) #PC1, PC2

Wmat <- list( W1=SimTable$Wmat )
names(Wmat) #W1

Ysim <- SimTable$Ysim

design_list <- vb_Designs(W=Wmat, X=Xmat, y=Ysim)
names(design_list$Y) # V1
#--------------------------------------------------------------------------

#specification of the prior distributions
beta_m<-matrix(0, nrow=3, ncol=1) #prior mean for beta
sigma_inv_beta<-diag(3)/1000 #prior covariance for beta

alpha_m <- matrix(0, nrow=2, ncol=1) #prior mean for alpha
sigma_inv_alpha <- diag(2)/1000 #prior covariance for alpha

# ~ occupancy ~ detection
#ndraws should be increased
occ <- PGocc4(formula= V1  ~ PC1 + PC2 ~ W1,
            design_mats = design_list,
            ndraws=1000,
            alpha_m=alpha_m, beta_m=beta_m,
            sigma_inv_alpha_p=sigma_inv_alpha,
            sigma_inv_beta_p=sigma_inv_beta,
            percent_burn_in=2/7)
#--------------------------------------------------------------------------

#not run
#require(coda)
#alpha_mcmc <- as.mcmc( t(spat_logit$alpha) ) #note the transpose
#summary(alpha_mcmc)
#effectiveSize(alpha_mcmc) #the effective sample size
#
#par(mfrow=c(2,2))
#traceplot(alpha_mcmc)
#densplot(alpha_mcmc)
#autocorr.plot(alpha_mcmc)
#run three chains to do convergence testing!
#--------------------------------------------------------------------------

attributes(occ)
dim(occ$alpha) #note the dimensions
dim(occ$beta) #note the dimensions

#some posterior stats
#--------------------------------------------------------------------------
apply(occ$alpha,1,mean)
apply(occ$alpha,1,sd)

apply(occ$beta,1,mean)
apply(occ$beta,1,sd)
#--------------------------------------------------------------------------

#Some trace plots
par(mfrow=c(2,2), mar=c(4,4,1,1))
plot(occ$alpha[1,], main="", xlab=expression(alpha[0]), type="l", ylab="")
plot(occ$alpha[2,], main="", xlab=expression(alpha[1]), type="l", ylab="")

plot(occ$beta[1,], main="", xlab=expression(beta[0]), type="l", col="red", ylab="")
plot(occ$beta[2,], main="", xlab=expression(beta[1]), type="l", col="red", ylab="")
#--------------------------------------------------------------------------

#plot of the posterior densities
par(mfrow=c(2,2), mar=c(4,4,1,1))
plot(density(occ$alpha[1,]), main="", xlab=expression(alpha[0]))
plot(density(occ$alpha[1,]), main="", xlab=expression(alpha[1]))

plot(density(occ$beta[1,]), main="", xlab=expression(beta[0]), col="red")
plot(density(occ$beta[1,]), main="", xlab=expression(beta[1]), col="red")
#--------------------------------------------------------------------------
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
