\name{vb_Designs_check}
\alias{vb_Designs_check}

\title{vb_Designs_check - An internal function.}

\description{
The function specifically checks that the names included in the formula call is the same as those included in the design matrices.}

\usage{
vb_Designs_check(formula, Names)
}
\arguments{
  \item{formula }{ - Double right-hand side formula describing covariates of occupancy and detection in that order.}
  \item{Names }{ - A named list generated by the call vb_Designs(W, X, y).}
}
