\name{Rcppocc-package}
\alias{Rcppocc-package}
\alias{Rcppocc}
\docType{package}

\title{Rcppocc}

\description{Fits various Bayesian occupancy models (nonspatial and spatial) assuming all regression effects uses a logit link function.}

\details{The following package allows the user to build various Bayesian
  occupancy models. At present the standard occupancy model as well a restricted
  spatial models can be undertaken. We specifically assume that the regression
  effects all use logit link functions. Gibbs sampling algorithms are developed
  to undertake all MCMC sampling.
}
\author{A.E. Clark (allan.clark@uct.ac.za)}

\references{
\itemize{
\item{Broms, K.M., Johnson, D.S., Altwegg, R. and Conquest, L.L., (2014). Spatial occupancy models applied to atlas data show Southern Ground Hornbills strongly depend on protected areas. Ecological Applications, 24(2), pp.363-374.}
\item Dorazio, R. M. and Rodriguez, D. T. (2012), A Gibbs sampler for Bayesian analysis of site-occupancy data. Methods in Ecology and Evolution, 3: 1093-1098. \url{doi: 10.1111/j.2041-210X.2012.00237.x}
\item Johnson, D. S., Conn, P. B., Hooten, M., Ray, J., and Pond, B. (2013). Spatial occupancy
models for large data sets. Ecology, \url{http://dx.doi.org/10.1890/12-0564.1.}
\item MacKenzie, D.I., Nichols, J.D., Lachman, G.B., Droege, S., Royle, J.A., Langtimm, C.A. (2002). Estimating site occupancy rates when detection probabilities are less than one. Ecology, 83(8): 2248-2255}
}

% Optionally other standard keywords, one per line, from file %
% KEYWORDS in the R documentation directory %

